requirements.txt
streamlit>=1.25.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
